+++ Westwind.Utilities/Utilities/ExensionMethods/MemoryStreamExtensions.cs
@@ -11,8 +11,8 @@ namespace System.IO
         /// <summary>
         /// Returns the content of the stream as a string
         /// </summary>
-        /// <param name="ms"></param>
-        /// <param name="encoding"></param>
+        /// <param name="ms">Memory stream</param>
+        /// <param name="encoding">Encoding to use - defaults to Unicode</param>
         /// <returns></returns>
         public static string AsString(this MemoryStream ms, Encoding encoding = null)
         {
@@ -35,6 +35,29 @@ namespace System.IO
 
             byte[] buffer = encoding.GetBytes(inputString);
             ms.Write(buffer, 0, buffer.Length);
+        }
+    }
+
+    /// <summary>
+    /// Stream Extensions
+    /// </summary>
+    public static class StreamExtensions
+    {
+        /// <summary>
+        /// Converts a stream by copying it to a memory stream and returning
+        /// as a string with encoding.
+        /// </summary>
+        /// <param name="s">stream to turn into a string</param>
+        /// <param name="encoding">Encoding of the stream. Defaults to Unicode</param>
+        /// <returns>string </returns>
+        public static string AsString(this Stream s, Encoding encoding = null)
+        {
+            using (var ms = new MemoryStream())
+            {
+                s.CopyTo(ms);
+                s.Position = 0;
+                return ms.AsString(encoding);
+            }
         }
     }
 }